REF: move arch in its own module.
