+++ okonomiyaki/platforms/tests/test_platform.py
@@ -5,124 +5,10 @@ from ..platform import Arch, Platform
 from ..platform import DARWIN, LINUX, MAC_OS_X, RHEL, WINDOWS, X86, X86_64
 
 from .common import (mock_machine_armv71, mock_x86, mock_x86_64,
-                     mock_x86_on_x86_64, mock_machine_x86_64)
+                     mock_machine_x86_64)
 from .common import (mock_architecture_64bit, mock_centos_3_5, mock_centos_5_8,
                      mock_centos_6_3, mock_centos_7_0, mock_osx_10_7,
                      mock_solaris, mock_ubuntu_raring, mock_windows_7)
-
-
-class TestArch(unittest.TestCase):
-    def test_simple(self):
-        # Given
-        name = "x86"
-        bits = 32
-
-        # When
-        arch = Arch(name, bits)
-
-        # Then
-        self.assertEqual(arch.name, name)
-        self.assertEqual(arch.bits, bits)
-        self.assertEqual(str(arch), arch.name)
-
-    def test_hashing(self):
-        # Given
-        name = "x86"
-
-        # When
-        arch1 = Arch(name, 32)
-        arch2 = Arch(name, 32)
-        arch3 = Arch(name, 64)
-
-        # Then
-        self.assertEqual(arch1, arch2)
-        self.assertNotEqual(arch1, arch3)
-        self.assertEqual(hash(arch1), hash(arch1))
-        self.assertTrue(arch1 == arch2)
-        self.assertFalse(arch1 != arch2)
-        self.assertTrue(arch1 != arch3)
-        self.assertFalse(arch1 == arch3)
-
-    def test_from_name(self):
-        # Given
-        name = "x86"
-
-        # When
-        arch = Arch.from_name(name)
-
-        # Then
-        self.assertEqual(arch.name, name)
-        self.assertEqual(arch.bits, 32)
-
-        # Given
-        name = "x86_64"
-
-        # When
-        arch = Arch.from_name(name)
-
-        # Then
-        self.assertEqual(arch.name, name)
-        self.assertEqual(arch.bits, 64)
-
-    def test_from_running_python(self):
-        # When
-        with mock_x86:
-            arch = Arch.from_running_python()
-
-        # Then
-        self.assertEqual(arch.name, "x86")
-        self.assertEqual(arch.bits, 32)
-
-        # When
-        with mock_x86_64:
-            arch = Arch.from_running_python()
-
-        # Then
-        self.assertEqual(arch.name, "x86_64")
-        self.assertEqual(arch.bits, 64)
-
-        # When
-        with mock_x86_on_x86_64:
-            arch = Arch.from_running_python()
-
-        # Then
-        self.assertEqual(arch.name, "x86")
-        self.assertEqual(arch.bits, 32)
-
-        # Given/When/Then
-        with mock_machine_armv71:
-            with self.assertRaises(OkonomiyakiError):
-                arch = Arch.from_running_python()
-
-    def test_from_running_system(self):
-        # When
-        with mock_x86:
-            arch = Arch.from_running_system()
-
-        # Then
-        self.assertEqual(arch.name, "x86")
-        self.assertEqual(arch.bits, 32)
-
-        # When
-        with mock_x86_64:
-            arch = Arch.from_running_system()
-
-        # Then
-        self.assertEqual(arch.name, "x86_64")
-        self.assertEqual(arch.bits, 64)
-
-        # When
-        with mock_x86_on_x86_64:
-            arch = Arch.from_running_system()
-
-        # Then
-        self.assertEqual(arch.name, "x86_64")
-        self.assertEqual(arch.bits, 64)
-
-        # Given/When/Then
-        with mock_machine_armv71:
-            with self.assertRaises(OkonomiyakiError):
-                arch = Arch.from_running_system()
 
 
 class TestPlatformRunningPython(unittest.TestCase):
